Kernel C++ support and build system cleanup
Set up tooling and basic runtime support for C++ in the kernel. "Basic
runtime support" means the few helper functions to get C++ to run
without the regular C++ runtime, since we don't need most of it (e.g.,
RTTI, exceptions). Note that the bootloader (and src/common) only
supports C code and the kernel only supports C++ code (at least due to
clangd needing to know if .h files are C or C++).

Also clean up .clangd and Makefile rules, add support for automatic
header dependencies, and add additional options to compile with the
GNU toolchain (gcc/g++/ld) and with different optimization levels.

+++ src/boot/page_table.h
@@ -105,7 +105,7 @@ struct page_directory_entry_4mb {
   bool pat : 1;
   // High bits of page address (e.g., for PSE-36). We're not going to
   // support that here, so this should always be zero.
-  uint8_t addr_ext : 8;
+  uint16_t addr_ext : 8;
   uint8_t rsv0 : 1;
   uint32_t addr : 10;
 } __attribute__((packed));
